pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "apify-client>=2.1.0",
    "bson>=0.5.10",
    "celery>=5.5.3",
    "chromadb>=1.1.1",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "mistralai>=1.9.11",
    "openai>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "ruff>=0.12.12",
    "scalar-fastapi>=1.3.0",
    "sqlmodel>=0.0.24",
    "tavily-python>=0.7.12",
    "uvicorn>=0.35.0",
]


[tool.ruff]
line-length = 120
lint.select = ["E", "F", "W", "I"]
lint.ignore = ["E722", "E741", "C901", "E501"]
target-version = "py312"
exclude = ["alembic/"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
